[refactor]: Set test targets
